Added authentication to meetings and registration controllers

+++ app/Http/Controllers/MeetingController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers;
 
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+
+use App\Http\Requests;
+
 use App\Meeting;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
     public function  __construct()
     {
-        //$this->middleware('name');
+        $this->middleware('jwt.auth', ['only' => [
+            'update', 'store', 'destroy'
+        ]]);
     }
 
     /**
@@ -48,14 +54,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+       if(!$user = JWTAuth::parseToken()->authenticate()) {
+           return response()->json(['msg' => 'User not found'], 404);
+       }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
        $meeting = new Meeting([
            'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -113,24 +122,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
-        $meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -141,6 +143,7 @@ class MeetingController extends Controller
         $meeting->time = Carbon::createFromFormat('YmdHie', $time);
         $meeting->title = $title;
         $meeting->description = $description;
+
         if(!$meeting->update()) {
             return response()->json(['msg' => 'Error during updating'], 404);
         }
@@ -167,6 +170,16 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+
+        if(!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        };
+
+
         $users = $meeting->users;
         $meeting->users()->detach();
         if(!$meeting->delete()) {
